got in-order insert working
